Moved AWS bucket selector UI elements to bisweb_awsmodule and improved the logic of logging into a new bucket
